ui: bugfix harbor transaction date info, withdraw request refactor
