+++ src/constants/config.ts
@@ -151,7 +151,7 @@ switch (Updates.channel) {
     break;
   }
   default: {
-    Config = envs.testnet;
+    Config = envs.prod;
     break;
   }
 }
